+++ react/package.json
@@ -7,6 +7,7 @@
     "express": "^4.16.4",
     "font-awesome": "^4.7.0",
     "json-server": "^0.14.2",
+    "npm": "^6.9.0",
     "react": "^16.8.5",
     "react-bootstrap": "^0.30.3",
     "react-dom": "^16.8.5",
