+++ react/src/App.js
@@ -14,8 +14,9 @@ import Spree from "./components/Game/SpreeGame/Game";
 import Infinite from "./components/Game/InfiniteGame/Game";
 import GameFinished from "./components/Game/ClassicGame/GameFinished";
 import Error from "./components/Error/Error";
-import ClassicMode from "./components/ClassicMode/ClassicMode";
-import ClassicGame from "./components/ClassicMode/ClassicGame";
+import ClassicMode from "./components/Modes/ClassicMode/ClassicMode";
+import SpreeMode from "./components/Modes/SpreeMode/SpreeMode";
+import InfiniteMode from "./components/Modes/InfiniteMode/InfiniteMode";
 import "./main.css";
 
 class App extends Component {
@@ -43,11 +44,12 @@ class App extends Component {
           <Route path="/signup" component={SignInPage} exact />
           <Route path="/gameselect" component={GameSelect} exact />
           <Route path="/classic" component={ClassicMode} exact />
+          <Route path="/spree" component={SpreeMode} exact />
+          <Route path="/infinite" component={InfiniteMode} exact />
           <Route path="/classic/game" component={Classic} exact />
           <Route path="/spree/game" component={Spree} exact />
           <Route path="/infinite/game" component={Infinite} exact />
           <Route path="/gamefinished" component={GameFinished} exact />
-          <Route path="/classicgame" component={ClassicGame} exact />
           <Route component={Error} />
         </Switch>
       </BrowserRouter>
