--- react/src/components/ClassicMode/classicgame.css
@@ -1,34 +1,0 @@
-.flexing {
-    display: flex;
-    justify-content: center;
-}
-
-.progress-barr {
-    height: 40px;
-    width: 25%;
-    background: repeating-linear-gradient(
-    to right,
-  #202225,
-  #202225 10px,
-  #36393f 10px,
-  #36393f 20px
-);
-}
-
-.fillerr {
-    background: linear-gradient(to right, rgb(255, 0, 0), rgb(230, 104, 0));
-    height: 100%;
-    border-radius:inherit;
-}
-
-.fillerr2 {
-    background: linear-gradient(to right, rgb(230, 104, 0), rgb(230, 226, 0));
-    height: 100%;
-    border-radius:inherit;
-}
-
-.fillerr3 {
-    background: linear-gradient(to right, rgb(230, 226, 0), rgb(73, 230, 0));
-    height: 100%;
-    border-radius:inherit;
-}
