+++ react/src/components/GameSelect/gameSelect.css
@@ -22,25 +22,25 @@
     font-size: 3rem;
 }
 
-.select-button {
-    margin-top: 10px;
+.game-button {
+    margin-top: 5px;
     width: 15rem;
-    background-color: #ff8648;
+    background-color: #ff7448;
     text-decoration: none;
     text-align: center;
     font-family: 'Montserrat';
     padding: 17px 0;
-    border: 1px solid #ff8648;
+    border: 1px solid #ff7448;
     border-radius: 3px;
     -webkit-transition-duration: 0.4s;
     transition-duration: 0.4s;
 }
 
-.select-button:hover {
+.game-button:hover {
     background-color: #36393f;
-    color: #ff8648;
+    color: #ff7448;
 }
 
-.select-button:focus {
+.game-button:focus {
     outline: none;
 }
