+++ react/src/components/Modes/modes.css
@@ -15,13 +15,22 @@
 .select-button {
     margin: 2.5px;
     width: 15rem;
-    background-color: #ff8648;
+    background-color: #ff7448;
     text-decoration: none;
     text-align: center;
     font-family: 'Montserrat';
     padding: 17px 0;
-    border: 1px solid #ff8648;
+    border: 1px solid #ff7448;
     border-radius: 3px;
     -webkit-transition-duration: 0.4s;
     transition-duration: 0.4s;
+}
+
+.select-button:hover {
+    background-color: #36393f;
+    color: #ff7448;
+}
+
+.game-button:focus {
+    outline: none;
 }
